requirements.txt
# Core data science libraries
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0

# Network analysis
networkx==3.2.1
scipy==1.11.4

# Text analysis and NLP
transformers==4.36.2
torch==2.1.2
nltk==3.8.1
textblob==0.17.1

# Data collection
requests==2.31.0
praw==7.7.1
python-dotenv==1.0.0
pushshift-py>=0.1.0

# Jupyter and visualization
jupyter>=1.0.0
plotly==5.17.0
wordcloud>=1.9.0

# Machine learning
scikit-learn==1.3.2

# Utilities
tqdm==4.66.1
python-dateutil==2.8.2 